Make Shade killable after talking to lamp post
